rename folder context -> contexts
